[REFACTOR] 라우터, 폴더 구조 변경
