+++ src/index.tsx
@@ -1,8 +1,17 @@
 import ReactDOM from "react-dom/client";
-import { RouterProvider } from "react-router-dom";
-import Router from "./Router";
+import { QueryClient, QueryClientProvider } from "react-query";
+import { ThemeProvider } from "styled-components";
+import App from "./App";
+import theme from "./styles/theme";
 
-const root = ReactDOM.createRoot(
-  document.getElementById("root") as HTMLElement
+const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement);
+
+const queryClient = new QueryClient();
+
+root.render(
+  <QueryClientProvider client={queryClient}>
+    <ThemeProvider theme={theme}>
+      <App />
+    </ThemeProvider>
+  </QueryClientProvider>
 );
-root.render(<RouterProvider router={Router} />);
